Externalized ts styles
